+++ skorch/net.py
@@ -34,6 +34,16 @@ from skorch.utils import to_numpy
 from skorch.utils import to_tensor
 
 
+_PYTORCH_COMPONENTS = {'criterion', 'module', 'optimizer'}
+"""Special names that mark pytorch components.
+
+These special names are used to recognize whether an attribute that is
+being set in the net should be added to prefixes_ and
+cuda_dependent_attributes_
+
+"""
+
+
 # pylint: disable=too-many-instance-attributes
 class NeuralNet:
     # pylint: disable=anomalous-backslash-in-string
@@ -403,7 +413,7 @@ class NeuralNet:
 
             # below: check for callback params
             # don't set a parameter for non-existing callback
-            params = self._get_params_for('callbacks__{}'.format(name))
+            params = self.get_params_for('callbacks__{}'.format(name))
             if (cb is None) and params:
                 raise ValueError("Trying to set a parameter for callback {} "
                                  "which does not exist.".format(name))
@@ -422,7 +432,7 @@ class NeuralNet:
 
     def initialize_criterion(self):
         """Initializes the criterion."""
-        criterion_params = self._get_params_for('criterion')
+        criterion_params = self.get_params_for('criterion')
         self.criterion_ = self.criterion(**criterion_params)
         if isinstance(self.criterion_, torch.nn.Module):
             self.criterion_ = to_device(self.criterion_, self.device)
@@ -452,7 +462,7 @@ class NeuralNet:
         reset.
 
         """
-        kwargs = self._get_params_for('module')
+        kwargs = self.get_params_for('module')
         module = self.module
         is_initialized = isinstance(module, torch.nn.Module)
 
@@ -506,7 +516,7 @@ class NeuralNet:
           about the parameters that caused the re-initialization.
 
         """
-        args, kwargs = self._get_params_for_optimizer(
+        args, kwargs = self.get_params_for_optimizer(
             'optimizer', self.module_.named_parameters())
 
         if self.initialized_ and self.verbose:
@@ -1141,7 +1151,7 @@ class NeuralNet:
         dataset = self.dataset
         is_initialized = not callable(dataset)
 
-        kwargs = self._get_params_for('dataset')
+        kwargs = self.get_params_for('dataset')
         if kwargs and is_initialized:
             raise TypeError("Trying to pass an initialized Dataset while "
                             "passing Dataset arguments ({}) is not "
@@ -1228,10 +1238,10 @@ class NeuralNet:
 
         """
         if training:
-            kwargs = self._get_params_for('iterator_train')
+            kwargs = self.get_params_for('iterator_train')
             iterator = self.iterator_train
         else:
-            kwargs = self._get_params_for('iterator_valid')
+            kwargs = self.get_params_for('iterator_valid')
             iterator = self.iterator_valid
 
         if 'batch_size' not in kwargs:
@@ -1245,19 +1255,48 @@ class NeuralNet:
     def _get_params_for(self, prefix):
         return params_for(prefix, self.__dict__)
 
-    def _get_params_for_optimizer(self, prefix, named_parameters):
-        """Parse kwargs configuration for the optimizer identified by
-        the given prefix. Supports param group assignment using wildcards:
+    def get_params_for(self, prefix):
+        """Collect and return init parameters for an attribute.
 
-            optimizer__lr=0.05,
-            optimizer__param_groups=[
-                ('rnn*.period', {'lr': 0.3, 'momentum': 0}),
-                ('rnn0', {'lr': 0.1}),
-            ]
+        Attributes could be, for instance, pytorch modules, criteria,
+        or data loaders (for optimizers, use
+        :meth:`.get_params_for_optimizer` instead). Use the returned
+        arguments to initialize the given attribute like this:
 
-        The first positional argument are the param groups.
+        .. code:: python
+
+            # inside initialize_module method
+            kwargs = self.get_params_for('module')
+            self.module_ = self.module(**kwargs)
+
+        Proceed analogously for the criterion etc.
+
+        The reason to use this method is so that it's possible to
+        change the init parameters with :meth:`.set_params`, which
+        in turn makes grid search and other similar things work.
+
+        Note that in general, as a user, you never have to deal with
+        this method because :meth:`.initialize_module` etc. are
+        already taking care of this. You only need to deal with this
+        if you override :meth:`.initialize_module` (or similar
+        methods) because you have some custom code that requires it.
+
+        Parameters
+        ----------
+        prefix : str
+          The name of the attribute whose arguments should be
+          returned. E.g. for the module, it should be ``'module'``.
+
+        Returns
+        -------
+        kwargs : dict
+          Keyword arguments to be used as init parameters.
+
         """
-        kwargs = self._get_params_for(prefix)
+        return self._get_params_for(prefix)
+
+    def _get_params_for_optimizer(self, prefix, named_parameters):
+        kwargs = self.get_params_for(prefix)
         params = list(named_parameters)
         pgroups = []
 
@@ -1273,15 +1312,67 @@ class NeuralNet:
 
         return [pgroups], kwargs
 
-    def _get_param_names(self):
-        return (k for k in self.__dict__.keys() if k != 'history_')
+    def get_params_for_optimizer(self, prefix, named_parameters):
+        """Collect and return init parameters for an optimizer.
 
-    def _get_param_names_new(self):
-        # TODO: This will be the new behavior for _get_param_names in
-        # a future release. This is to make get_params work as in
-        # sklearn, i.e. not returning "learned" attributes (ending on
-        # '_'). Once the transition period has passed, remove the old
-        # code and use the new one instead.
+        Parse kwargs configuration for the optimizer identified by
+        the given prefix. Supports param group assignment using wildcards:
+
+        .. code:: python
+
+            optimizer__lr=0.05,
+            optimizer__param_groups=[
+                ('rnn*.period', {'lr': 0.3, 'momentum': 0}),
+                ('rnn0', {'lr': 0.1}),
+            ]
+
+        Generally, use this method like this:
+
+        .. code:: python
+
+            # inside initialize_optimizer method
+            named_params = self.module_.named_parameters()
+            pgroups, kwargs = self.get_params_for_optimizer('optimizer', named_params)
+            if 'lr' not in kwargs:
+                kwargs['lr'] = self.lr
+            self.optimizer_ = self.optimizer(*pgroups, **kwargs)
+
+        The reason to use this method is so that it's possible to
+        change the init parameters with :meth:`.set_params`, which
+        in turn makes grid search and other similar things work.
+
+        Note that in general, as a user, you never have to deal with
+        this method because :meth:`.initialize_optimizer` is already
+        taking care of this. You only need to deal with this if you
+        override :meth:`.initialize_optimizer` because you have some
+        custom code that requires it.
+
+        Parameters
+        ----------
+        prefix : str
+          The name of the optimizer whose arguments should be
+          returned. Typically, this should just be
+          ``'optimizer'``. There can be exceptions, however, e.g. if
+          you want to use more than one optimizer.
+
+        named_parameters : iterator
+          Iterator over the parameters of the module that is intended
+          to be optimized. It's the return value of
+          ``my_module.named_parameters()``.
+
+        Returns
+        -------
+        pgroups : list
+          List of parameter groups.
+
+        kwargs : dict
+          All other parameters for this optimizer, e.g. the learning
+          rate.
+
+        """
+        return self._get_params_for_optimizer(prefix, named_parameters)
+
+    def _get_param_names(self):
         return (k for k in self.__dict__ if not k.endswith('_'))
 
     def _get_params_callbacks(self, deep=True):
@@ -1355,7 +1446,7 @@ class NeuralNet:
             tmpl = ("__init__() got unexpected argument(s) {}. "
                     "Either you made a typo, or you added new arguments "
                     "in a subclass; if that is the case, the subclass "
-                    "should deal with the new arguments explicitely.")
+                    "should deal with the new arguments explicitly.")
             msg = tmpl.format(', '.join(sorted(unexpected_kwargs)))
             msgs.append(msg)
 
@@ -1416,17 +1507,17 @@ class NeuralNet:
             self.initialize_callbacks()
             self._set_params_callback(**cb_params)
 
-        if any(key.startswith('criterion') for key in special_params):
+        if any('criterion' in key.split('__', 1)[0] for key in special_params):
             self.initialize_criterion()
 
         module_triggers_optimizer_reinit = False
-        if any(key.startswith('module') for key in special_params):
+        if any('module' in key.split('__', 1)[0] for key in special_params):
             self.initialize_module()
             module_triggers_optimizer_reinit = True
 
         optimizer_changed = (
-            any(key.startswith('optimizer') for key in special_params) or
-            'lr' in normal_params
+            any('optimizer' in key.split('__', 1)[0] for key in special_params)
+            or 'lr' in normal_params
         )
         if module_triggers_optimizer_reinit or optimizer_changed:
             # Model selectors such as GridSearchCV will set the
@@ -1517,6 +1608,100 @@ class NeuralNet:
         state.pop('__cuda_dependent_attributes__')
 
         self.__dict__.update(state)
+
+    def _register_attribute(
+            self,
+            name,
+            prefixes=True,
+            cuda_dependent_attributes=True,
+    ):
+        """Add attribute name to prefixes_ and
+        cuda_dependent_attributes_.
+
+        The first is to take care that the attribute works correctly
+        with set_params, e.g. when it comes to re-initialization.
+
+        The second is to make sure that nets trained with CUDA can be
+        loaded without CUDA.
+
+        This method takes care of not mutating the lists.
+
+        Parameters
+        ----------
+        prefixes : bool (default=True)
+          Whether to add to prefixes_.
+
+        cuda_dependent_attributes : bool (default=True)
+          Whether to add to cuda_dependent_attributes_.
+
+        """
+        # copy the lists to avoid mutation
+        if prefixes:
+            self.prefixes_ = self.prefixes_[:] + [name]
+
+        if cuda_dependent_attributes:
+            self.cuda_dependent_attributes_ = (
+                self.cuda_dependent_attributes_[:] + [name + '_'])
+
+    def _unregister_attribute(
+            self,
+            name,
+            prefixes=True,
+            cuda_dependent_attributes=True,
+    ):
+        """Remove attribute name from prefixes_ and
+        cuda_dependent_attributes_.
+
+        Use this to remove PyTorch components that are not needed
+        anymore. This is mostly a clean up job, so as to not leave
+        unnecessary prefixes or cuda-dependent attributes.
+
+        This method takes care of not mutating the lists.
+
+        Parameters
+        ----------
+        prefixes : bool (default=True)
+          Whether to remove from prefixes_.
+
+        cuda_dependent_attributes : bool (default=True)
+          Whether to remove from cuda_dependent_attributes_.
+
+        """
+        # copy the lists to avoid mutation
+        if prefixes:
+            self.prefixes_ = [p for p in self.prefixes_ if p != name]
+
+        if cuda_dependent_attributes:
+            self.cuda_dependent_attributes_ = [
+                a for a in self.cuda_dependent_attributes_ if a != name + '_']
+
+    def __setattr__(self, name, attr):
+        """Set an attribute on the net
+
+        When a custom net with additional torch modules or optimizers
+        is created, those attributes are added to ``prefixes_`` and
+        ``cuda_dependent_attributes_`` automatically.
+
+        """
+        # If it's a
+        # 1. known attribute or
+        # 2. special param like module__num_units or
+        # 3. not a torch module/optimizer instance or class
+        # just setattr as usual.
+        # For a discussion why we chose this implementation, see here:
+        # https://github.com/skorch-dev/skorch/pull/597
+        is_known = name.endswith('_') or (name in self.prefixes_)
+        is_special_param = '__' in name
+        is_torch_component = any(c in name for c in _PYTORCH_COMPONENTS)
+
+        if not (is_known or is_special_param) and is_torch_component:
+            self._register_attribute(name)
+        super().__setattr__(name, attr)
+
+    def __delattr__(self, name):
+        # take extra precautions to undo the changes made in __setattr__
+        self._unregister_attribute(name)
+        super().__delattr__(name)
 
     def save_params(
             self, f_params=None, f_optimizer=None, f_history=None):